+++ wezterm-gui/src/overlay/mod.rs
@@ -12,6 +12,7 @@ pub mod debug;
 pub mod launcher;
 pub mod prompt;
 pub mod quickselect;
+pub mod selector;
 
 pub use confirm_close_pane::{
     confirm_close_pane, confirm_close_tab, confirm_close_window, confirm_quit_program,